flask_test/requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
pyyaml
pymysql
flask_httpauth
